requirements.txt
Django~=4.1.3
djangorestframework~=3.14.0
pygments~=2.13.0
psycopg2-binary~=2.9.5
loguru~=0.6.0
aiormq~=6.4.2
pydantic~=1.10.2
python-dotenv~=0.21.0
pika~=1.3.1
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
PyJWT~=2.6.0